+++ HISTORY.rst
@@ -24,6 +24,7 @@ Latest (Unreleased)
 * BSIBSO-1093: Install module bso_telephony
 * BSIBSO-1093: Replace Dial button by a phone icon
 * purchase order: add a cron to recompute the received qty each day
+* update bso_backbone: cympa sychronization
 
 **Bugfixes**
 * purchase order invoicing: fix the received quantity computation